// src/app/features/dashboard/model/cases.model.ts
export interface Cases {
  confirmed: number;
  recovered: number;
  deaths: number;
  country: string;
  population: number;
  sq_km_area: number;
  life_expectancy: string;
  elevation_in_meters: number;
  continent: string;
  abbreviation: string;
  location: string;
  iso: string;
  capital_city: string;
  lat: string;
  long: string;
  updated: Date;
}