// src/app/features/dashboard/model/vaccines.model.ts
export interface Vaccines {
  administered: number;
  people_vaccinated: number;
  country: string;
  population: number;
  sq_km_area: number;
  life_expectancy: string;
  elevation_in_meters: number;
  continent: string;
  abbreviation: string;
  location: string;
  iso: string;
  capital_city: string;
  lat: string;
  long: string;
  updated: Date;
}